perf(remapping): make render chunk sourcemap encode mapping in parallel
